+++ android/quest/src/main/java/org/smartregister/fhircore/quest/ui/appsetting/AppSettingViewModel.kt
@@ -27,13 +27,16 @@ import java.nio.charset.StandardCharsets
 import javax.inject.Inject
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
+import okhttp3.RequestBody.Companion.toRequestBody
 import okio.ByteString.Companion.decodeBase64
+import org.apache.commons.lang3.StringUtils
 import org.hl7.fhir.r4.model.Binary
+import org.hl7.fhir.r4.model.Bundle
 import org.hl7.fhir.r4.model.Composition
 import org.hl7.fhir.r4.model.ResourceType
+import org.jetbrains.annotations.VisibleForTesting
 import org.smartregister.fhircore.engine.BuildConfig
 import org.smartregister.fhircore.engine.R
-import org.smartregister.fhircore.engine.configuration.ConfigType
 import org.smartregister.fhircore.engine.configuration.ConfigurationRegistry
 import org.smartregister.fhircore.engine.configuration.ConfigurationRegistry.Companion.DEBUG_SUFFIX
 import org.smartregister.fhircore.engine.configuration.app.ConfigService
@@ -41,11 +44,13 @@ import org.smartregister.fhircore.engine.configuration.profile.ProfileConfigurat
 import org.smartregister.fhircore.engine.configuration.register.RegisterConfiguration
 import org.smartregister.fhircore.engine.data.local.DefaultRepository
 import org.smartregister.fhircore.engine.data.remote.fhir.resource.FhirResourceDataSource
+import org.smartregister.fhircore.engine.di.NetworkModule
 import org.smartregister.fhircore.engine.domain.model.FhirResourceConfig
 import org.smartregister.fhircore.engine.domain.model.ResourceConfig
 import org.smartregister.fhircore.engine.util.DispatcherProvider
 import org.smartregister.fhircore.engine.util.SharedPreferenceKey
 import org.smartregister.fhircore.engine.util.SharedPreferencesHelper
+import org.smartregister.fhircore.engine.util.extension.encodeResourceToString
 import org.smartregister.fhircore.engine.util.extension.extractId
 import org.smartregister.fhircore.engine.util.extension.getActivity
 import org.smartregister.fhircore.engine.util.extension.launchActivityWithNoBackStackHistory
@@ -66,6 +71,8 @@ constructor(
   val configurationRegistry: ConfigurationRegistry,
   val dispatcherProvider: DispatcherProvider,
 ) : ViewModel() {
+
+  private var _isNonProxy = BuildConfig.IS_NON_PROXY_APK
 
   val showProgressBar = MutableLiveData(false)
 
@@ -121,32 +128,50 @@ constructor(
             val chunkedResourceIdList =
               entry.value.chunked(ConfigurationRegistry.MANIFEST_PROCESSOR_BATCH_SIZE)
             chunkedResourceIdList.forEach { parentIt ->
-              val ids = parentIt.joinToString(",") { it.focus.extractId() }
-              val resourceUrlPath =
-                "${entry.key}?${Composition.SP_RES_ID}=$ids&_count=${ConfigurationRegistry.DEFAULT_COUNT}"
-              Timber.d("Fetching config: $resourceUrlPath")
-              fhirResourceDataSource.getResource(resourceUrlPath).entry.forEach {
-                bundleEntryComponent ->
-                defaultRepository.createRemote(false, bundleEntryComponent.resource)
+              Timber.d(
+                "Fetching config resource ${entry.key}: with ids ${StringUtils.join(parentIt,",")}",
+              )
 
-                if (bundleEntryComponent.resource is Binary) {
-                  val binary = bundleEntryComponent.resource as Binary
-                  binary.data.decodeToString().decodeBase64()?.string(StandardCharsets.UTF_8)?.let {
-                    val registerConfig = it.tryDecodeJson<RegisterConfiguration>()
-                    if (registerConfig != null) {
-                      if (registerConfig.configType == ConfigType.Profile.name) {
-                        val profileConfig = it.tryDecodeJson<ProfileConfiguration>()
-                        profileConfig
-                          ?.fhirResource
-                          ?.dependentResourceTypes(
-                            patientRelatedResourceTypes,
-                          )
-                      } else {
-                        registerConfig.fhirResource.dependentResourceTypes(
-                          patientRelatedResourceTypes,
-                        )
+              val resultBundle: Bundle =
+                if (isNonProxy()) {
+                  fhirResourceDataSourceGetBundle(
+                    entry.key,
+                    parentIt.map { it.focus.extractId() },
+                  )
+                } else
+                  fhirResourceDataSource.post(
+                    requestBody =
+                      generateRequestBundle(entry.key, parentIt.map { it.focus.extractId() })
+                        .encodeResourceToString()
+                        .toRequestBody(NetworkModule.JSON_MEDIA_TYPE),
+                  )
+
+              resultBundle.entry.forEach { bundleEntryComponent ->
+                if (bundleEntryComponent.resource != null) {
+                  defaultRepository.createRemote(false, bundleEntryComponent.resource)
+
+                  if (bundleEntryComponent.resource is Binary) {
+                    val binary = bundleEntryComponent.resource as Binary
+                    binary.data
+                      .decodeToString()
+                      .decodeBase64()
+                      ?.string(StandardCharsets.UTF_8)
+                      ?.let {
+                        val config =
+                          it.tryDecodeJson<RegisterConfiguration>()
+                            ?: it.tryDecodeJson<ProfileConfiguration>()
+
+                        when (config) {
+                          is RegisterConfiguration ->
+                            config.fhirResource.dependentResourceTypes(
+                              patientRelatedResourceTypes,
+                            )
+                          is ProfileConfiguration ->
+                            config.fhirResource.dependentResourceTypes(
+                              patientRelatedResourceTypes,
+                            )
+                        }
                       }
-                    }
                   }
                 }
               }
@@ -195,7 +220,7 @@ constructor(
             sharedPreferencesHelper.write(SharedPreferenceKey.APP_ID.name, thisAppId)
             context.getActivity()?.launchActivityWithNoBackStackHistory<LoginActivity>()
           } else {
-            _error.postValue(context.getString(R.string.application_not_supported, appId.value))
+            _error.postValue(context.getString(R.string.application_not_supported, thisAppId))
           }
         }
       }
@@ -219,5 +244,58 @@ constructor(
   }
 
   fun hasDebugSuffix(): Boolean =
-    appId.value?.endsWith(DEBUG_SUFFIX, ignoreCase = true) == true && BuildConfig.DEBUG
+    appId.value?.endsWith(DEBUG_SUFFIX, ignoreCase = true) == true && isDebugVariant()
+
+  @VisibleForTesting fun isDebugVariant() = BuildConfig.DEBUG
+
+  private fun generateRequestBundle(resourceType: String, idList: List<String>): Bundle {
+    val bundleEntryComponents = mutableListOf<Bundle.BundleEntryComponent>()
+
+    idList.forEach {
+      bundleEntryComponents.add(
+        Bundle.BundleEntryComponent().apply {
+          request =
+            Bundle.BundleEntryRequestComponent().apply {
+              url = "$resourceType/$it"
+              method = Bundle.HTTPVerb.GET
+            }
+        },
+      )
+    }
+
+    return Bundle().apply {
+      type = Bundle.BundleType.BATCH
+      entry = bundleEntryComponents
+    }
+  }
+
+  private suspend fun fhirResourceDataSourceGetBundle(
+    resourceType: String,
+    resourceIds: List<String>,
+  ): Bundle {
+    val bundleEntryComponents = mutableListOf<Bundle.BundleEntryComponent>()
+
+    resourceIds.forEach {
+      val responseBundle =
+        fhirResourceDataSource.getResource("$resourceType?${Composition.SP_RES_ID}=$it")
+      responseBundle?.let {
+        bundleEntryComponents.add(
+          Bundle.BundleEntryComponent().apply {
+            resource = responseBundle.entry?.first()?.resource
+          },
+        )
+      }
+    }
+    return Bundle().apply {
+      type = Bundle.BundleType.COLLECTION
+      entry = bundleEntryComponents
+    }
+  }
+
+  @VisibleForTesting fun isNonProxy(): Boolean = _isNonProxy
+
+  @VisibleForTesting
+  fun setNonProxy(nonProxy: Boolean) {
+    _isNonProxy = nonProxy
+  }
 }