+++ android/quest/src/main/java/org/smartregister/fhircore/quest/ui/report/measure/MeasureReportMainScreen.kt
@@ -41,6 +41,7 @@ fun MeasureReportMainScreen(
 ) {
   // Use a different navController internally for navigating Report Composable screens
   val navController = rememberNavController()
+  val uiState = measureReportViewModel.reportTypeSelectorUiState.value
 
   NavHost(
     navController = navController,
@@ -60,6 +61,7 @@ fun MeasureReportMainScreen(
             ),
           )
         },
+        showProgressIndicator = uiState.showProgressIndicator,
       )
     }
     // Page for selecting report date
@@ -76,7 +78,7 @@ fun MeasureReportMainScreen(
       ReportDateSelectorScreen(
         reportId = reportId,
         practitionerId = practitionerId,
-        screenTitle = stringResource(R.string.select_date_range),
+        screenTitle = stringResource(R.string.select_month),
         navController = navController,
         mainNavController = mainNavController,
         measureReportViewModel = measureReportViewModel,