+++ android/quest/src/test/java/org/smartregister/fhircore/quest/ui/report/measure/MeasureReportViewModelTest.kt
@@ -291,7 +291,7 @@ class MeasureReportViewModelTest : RobolectricTest() {
 
     assertEquals(1, result.size)
     assertEquals("report group 1", result.first().title)
-    assertEquals(0, result.first().dataList.size)
+    assertEquals(1, result.first().dataList.size)
   }
 
   private val measureReport =
@@ -371,10 +371,10 @@ class MeasureReportViewModelTest : RobolectricTest() {
     assertEquals("", result.first().count)
     assertEquals(0, result.first().dataList.size)
     assertEquals("Test Code", result.first().indicatorTitle)
-    assertEquals(3, result.first().measureReportDenominator)
+    assertEquals("3", result.first().measureReportDenominator)
 
     assertEquals("", result.last().count)
-    assertEquals(0, result.last().dataList.size)
+    assertEquals(1, result.last().dataList.size)
   }
 
   @Test
@@ -413,7 +413,7 @@ class MeasureReportViewModelTest : RobolectricTest() {
     assertEquals("", result.first().count)
     assertEquals(3, result.first().dataList.size)
     assertEquals("Commodity 1", result.first().indicatorTitle)
-    assertEquals(null, result.first().measureReportDenominator)
+    assertEquals("0", result.first().measureReportDenominator)
 
     assertEquals("Test Code 1", result.first().dataList.elementAt(0).title)
     assertEquals("2", result.first().dataList.elementAt(0).count)