+++ CHANGELOG.md
@@ -10,10 +10,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Insights feature to show stats on any _Unsynced_ Resources on the device
 - Generic date service function to add or subtract days, weeks, months or years from/to current date
 - Add ability to configure resource Id used to open profiles
+- [App Performance] Profile the time it takes for the app to load the configs and patient data from the local DB
+- Add progress indicator during report generation
 
 ### Fixed
 - Incorrect error shown for failed authentication when the credentials are invalid.
 - [P2P] Possible p2p sync slow regression #2536
+- [Event Management] Fix resource closure failure after server sync
 
 ### Changed
 - 
@@ -52,7 +55,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Implement configurable image/icon widget
 - Added functionality to filter related resources for list view
 - Added the sick child workflow automated closure
+<<<<<<< HEAD
 - Added a new eCBIS flavour to be used for the production app.
+=======
+- Added a new eCBIS flavour to be used for the production app. 
+>>>>>>> 079b25a2e538573a891985a1de771734283037e5
 
 ### Fixed
 - Logout while offline causes flicker(multiple render) of Login Page